requirements.txt
streamlit==1.55.0
pandas==2.3.3
requests==2.32.5
openpyxl==3.1.5
xlrd==2.0.2
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
